:recycle: Refactor spotify module
- Rename SpotifyPlaylistRetrievalClient#fetch() argument: queries -> params
- Move SpotifyClientCredentials from 'model' to 'credential' package
